requirements.txt
openpyxl==3.1.5
docxtpl==0.20.2
python-docx==1.2.0
customtkinter==5.2.2
